cpufreq: times: add /proc/uid_cpupower/concurrent_policy_time
In order to model the energy used by the cpu,
we need to expose cpu time information to
userspace. Output the time each uid spent running
on group of cpu's with the same policy with 1
to num_related_cpus online and not idle.

The time should also include the time spent by
pids that are no longer alive.

We can use this information to blame uids for
the energy they are using.

For instance, if 2 cores that are a part of the
same policy are online and active for 50ms and
the uid is running on one of the cores, the uid
should be blamed for 1/2 of the energy that is
used everytime one or more cpus in the same policy
is active. Using other proc files, the uid can
also be blamed for the cost of keeping the entire
cpu unsuspended and the cost of the individual
cores it is using.

This patch is based on commit
9892125368426521075d79c464123076213ea3b7 in the
android-msm-wahoo-4.4 branch by Marissa Wall.

Bug: 68399339
Bug: 74242328
Test: read file, check that values are reasonable
Change-Id: Ibfaccd5b5b249de7420952d83bdf7d1c12311d4f
Signed-off-by: Connor O'Brien <connoro@google.com>
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>
Signed-off-by: Abhishek AN <darkabhi1520@gmail.com>